cleaning up

+++ src/entities/enemy.rs
@@ -65,8 +65,8 @@ pub enum EnemyState {
 }
 
 pub struct EnemyData {
-    pub health: f64,
-    pub shield_power: f64,
+    pub health: i32,
+    pub shield_power: i32,
     pub pos: Vector2,
     pub state: EnemyState,
 
@@ -96,7 +96,7 @@ impl Enemy {
         }
     }
 
-    pub fn take_hit(&mut self, damage: f64) {
+    pub fn take_hit(&mut self, damage: i32) {
         self.get_mut_data().health -= damage;
         self.get_mut_data().state = EnemyState::HitDelay;
     }
